SFR-2262: Refactoring cluster process
